Add initial warning and confirmation to setup.sh

+++ setup.sh
@@ -15,6 +15,17 @@ OptionalCopy() {
     fi
 }
 
+# Warn the user before doing anything
+echo "WARNING: This will overwrite your ~/.vimrc and ~/.bash_aliases files automatically, with the option to overwrite your ~/.bashrc and ~/.gitconfig files"
+read -p "Would you like to continue? (Y/n) " continue_var
+
+if [[ ! $continue_var =~ $yn_regex ]] # If not continue, quit script
+then
+    exit 0 # Exit successfully
+fi
+
+echo
+
 # Copy the .vimrc and .bash_aliases to the home directory
 cp .vimrc ~
 echo ".vimrc copied"
@@ -35,3 +46,5 @@ fi
 echo
 
 OptionalCopy .gitconfig
+
+exit 0 # Exit successfully
